requirements.txt
python-spreadsheet
django-dept-directory
